move gevent to the gcal client lib
